+++ src/main/java/rabun/oanda/rest/models/PositionClosed.java
@@ -1,7 +1,7 @@
 package rabun.oanda.rest.models;
 
 public class PositionClosed {
-    public int[] ids;
+    public long[] ids;
     public String instrument;
     public long totalUnits;
     public float price;
